\name{assocInd-package}
\alias{assocInd-package}
\alias{assocInd}
\docType{package}
\title{
Implements New and Existing Association Indices for Constructing Animal Social Networks
~~ assocInd ~~
}
\description{
Implements several new association indices that can control for various types of errors. Also includes existing association indices and functions for simulating the effects of different rates of error on estimates of association strength between individuals using each method.
}
\details{
\tabular{ll}{
Package: \tab assocInd\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2017-07-05\cr
License: \tab GPL2\cr
}
}
\author{
Authors:
	William Hoppitt <W.J.E.Hoppitt@leeds.ac.uk>
	Damien Farine <dfarine@orn.mpg.de>

Maintainer: Damien Farine  <dfarine@orn.mpg.de>
}
\references{
	Hoppitt, W. & Farine, D.R. (in press) Association indices for quantifying social relationships: how to deal with missing observations of individuals or groups. Animal Behaviour
}
\keyword{ package }
